+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## next
 
 - Added `detailsWhen` option for `table` cells to explicitly control the activation of details
+- Changed `button` view to take `text`, `href` and `external` options; it attempts to derive these values from `data` when options are not explicitly provided for backward compatibility
+- Changed `button` view so that `disabled` values are no longer treated as queries. To use a query for disabling a button, specify it explicitly like `{ view: "button", disabled: "=query" }`.
 - Fixed TypeScript warning related to `switch` view export due `switch` is reserved word and can't be an identifier
 
 ## 1.0.0-beta.87 (15-10-2024)